task

+++ python_async_comprehension/1-async_comprehension.py
@@ -9,7 +9,7 @@ import importlib
 
 async_generator = importlib.import_module("0-async_generator").async_generator
 
-async def async_comprehension() -> List[int]:
+async def async_comprehension() -> List[float]:
     """
     Collects 10 random numbers using an async comprehension over async_generator.
